Update the type of Role ID.
Repository.findOne cannot deal with int type. Change it to Long.

+++ Bookstore/src/main/java/com/bookstore/controller/HomeController.java
@@ -127,7 +127,7 @@ public class HomeController {
 		user.setPassword(encryptedPassword);
 		
 		Role role = new Role();
-		role.setId(1);
+		role.setId((long) 1);
 		role.setName("ROLE_USER");
 		Set<UserRole> userRoles = new HashSet<>();
 		userRoles.add(new UserRole(user, role));
@@ -627,7 +627,7 @@ public class HomeController {
 		user.setPassword(encryptedPassword);
 		
 		Role role = new Role();
-		role.setId(1);
+		role.setId((long)1);
 		role.setName("ROLE_USER");
 		Set<UserRole> userRoles = new HashSet<>();
 		userRoles.add(new UserRole(user, role));
